create run_seed_data script and cleanup seeding
